Rewrote how MySQL class works, making things much simpler and easier to understand.
MySQL class now works more like I would like it to. You make a query and it creates a 2d array based on the columns and rows that are returned. This allows for much easier access to a specific value in the database. You can run a query and access it like so:
    MySQL_Result res;
    value = res[2]["MyField"];
Where `res[2]' is the 3rd row returned from the result. (Note you still must run MySQL_Query to get a MySQL_Result object)